migration fixes

+++ src/airunner/alembic/versions/536e18463461_create_whisper_settings_table.py
@@ -5,16 +5,13 @@ Revises: 3212e8fccc68
 Create Date: 2024-10-13 09:34:20.974389
 
 """
-from typing import Sequence, Union
+from typing import Union
 from alembic import op
-from sqlalchemy.engine.reflection import Inspector
+from airunner.utils.db.table import table_exists
 from airunner.data.models import WhisperSettings
 
-# revision identifiers, used by Alembic.
 revision: str = '536e18463461'
 down_revision: Union[str, None] = '3212e8fccc68'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
 
 def set_default_values(model_name_):
     default_values = {}
@@ -27,10 +24,7 @@ def set_default_values(model_name_):
     )
 
 def upgrade():
-    inspector = Inspector.from_engine(op.get_bind())
-    tables = inspector.get_table_names()
-
-    if WhisperSettings.__tablename__ not in tables:
+    if not table_exists(WhisperSettings.__tablename__):
         try:
             op.create_table(
                 WhisperSettings.__tablename__,
@@ -47,8 +41,8 @@ def upgrade():
         print("whisper_settings already exists, skipping")
 
 def downgrade():
-    try:
-        op.drop_table('whisper_settings')
-    except Exception as e:
-        print("Failed to drop table", str(e))
-    # ### end Alembic commands ###
+    if table_exists(WhisperSettings.__tablename__):
+        try:
+            op.drop_table('whisper_settings')
+        except Exception as e:
+            print("Failed to drop table", str(e))
